all in one sell button with approve & setPrice
