+++ common/lib/kconfig.ts
@@ -285,8 +285,8 @@ const storageLinkToUrlHttp = (link: string): string =>
   link.startsWith("ipfs://") || link.startsWith(IPFS_GATEWAY)
     ? ipfsLinkToUrlHttp(link)
     : link.startsWith("swarm://") || link.startsWith(SWARM_GATEWAY)
-      ? swarmLinkToUrlHttp(link)
-      : link;
+    ? swarmLinkToUrlHttp(link)
+    : link;
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -474,6 +474,7 @@ export {
   explorerCollectionUrl,
   explorerCollectionLink,
   kredeumNftUrl,
+  explorerUrl,
   explorerNftUrl,
   explorerAccountUrl,
   explorerNftLink,
