+++ svelte/components/Nft/Nft.svelte
@@ -5,14 +5,12 @@
 
   import type { NftType } from "@lib/ktypes";
   import {
-    nftUrl,
     explorerCollectionUrl,
     explorerAddressLink,
     kredeumNftUrl,
     getOpenSea,
     nftOpenSeaUrl,
-    addressSame,
-    sleep
+    addressSame
   } from "@lib/kconfig";
 
   import MediaPreview from "../Media/MediaPreview.svelte";
@@ -24,14 +22,13 @@
   import NftBuy from "./NftBuy.svelte";
   import NftBurn from "./NftBurn.svelte";
 
-  import NftSetPrice from "./NftSetPrice.svelte";
-  import { setTokenPrice } from "@lib/kautomarket";
+  import NftSell from "./NftSell.svelte";
+  import { getEthersConverterLink } from "@lib/kautomarket";
 
   // import NftClaim from "./NftClaim.svelte";
 
-  import { metamaskChainId, metamaskSigner, metamaskProvider } from "@main/metamask";
-  import { collectionStore } from "@stores/collection/collection";
-  import { collectionGetContract, collectionGet } from "@lib/kcollection-get";
+  import { metamaskChainId, metamaskProvider } from "@main/metamask";
+  import { collectionGetContract } from "@lib/kcollection-get";
 
   /////////////////////////////////////////////////
   //  <Nft {chainId} {address} {tokenID} {account}? {platform}? />
@@ -47,15 +44,9 @@
 
   let burnable: boolean = false;
 
-  let setPriceInput: string;
-  let newNftPrice: string;
-  let settingTokenPrice: boolean = false;
-
-  let refreshCollectionList: Writable<number> = getContext("refreshCollectionList");
-
   // let i = 1;
   // HANDLE CHANGE : on truthy chainId and address, and whatever account
-  $: account, $refreshCollectionList && chainId && address && tokenID && $metamaskChainId && handleChange();
+  $: account, chainId && address && tokenID && $metamaskChainId && handleChange();
   const handleChange = async (): Promise<void> => {
     // console.log(`NFTDETAIL CHANGE #${i++} ${nftKey(chainId, address, tokenID)}`);
 
@@ -66,64 +57,11 @@
     await nftStore.refreshOne(chainId, address, tokenID).catch(console.error);
   };
 
-  $: console.log("Nft", $nft);
-  $: console.log("Nft price ", $nft.price);
-
   $: chainId && address && account && checkBurnable();
   const checkBurnable = async () => {
     const { contract, supports } = await collectionGetContract(chainId, address, $metamaskProvider);
-    // console.log("🚀 ~ file: Nft.svelte ~ line 71 ~ checkBurnable ~ supports", contract);
-    // const collection = await collectionGet(chainId, address, $metamaskProvider);
-
-    console.log("🚀 ~ file: Nft.svelte ~ line 73 ~ checkBurnable ~ collection", supports);
 
     burnable = supports?.IOpenNFTsV4;
-  };
-
-  /////////////////////////////////////////////////
-  $: setPriceInput && handlePrice();
-  const handlePrice = () => {
-    setPriceInput = setPriceInput.replace(/[^0-9.,]/g, "");
-    let formatedInputPrice = setPriceInput.replace(/[,]/g, ".");
-    const decimals = formatedInputPrice.split(".")[1];
-    if (decimals?.length > 18) {
-      setPriceInput = setPriceInput.slice(0, -1);
-      formatedInputPrice = formatedInputPrice.slice(0, -1);
-    }
-
-    // const priceToConvert = inputPrice.replace(/[,]/g, ".").replace(/[^0-9.]/g, "");
-    if (setPriceInput) newNftPrice = formatedInputPrice;
-
-    console.log("nftPrice : ", newNftPrice, " Wei");
-  };
-
-  const displayPriceInput = () => {
-    setPriceInput = $nft?.price;
-  };
-
-  const setNewTokenPrice = async () => {
-    settingTokenPrice = true;
-    const tansReceipt = await setTokenPrice(chainId, address, $metamaskSigner, tokenID, newNftPrice);
-
-    const blockTx = tansReceipt.blockNumber;
-
-    // do {
-    //   await sleep(1000);
-    // } while ((await $metamaskProvider.getBlockNumber()) <= blockTx);
-
-    // await nftStore.refreshOne(chainId, address, tokenID).catch(console.error);
-    // nft = nftStore.getOneStore(chainId, address, tokenID);
-
-    // const { contract, supports } = await collectionGetContract(chainId, address, $metamaskProvider);
-    // if (supports.IOpenMarketable) {
-    //   const openNFTsV4 = contract as OpenNFTsV4;
-    //   $nft.price = ethers.utils.formatEther(
-    //     (await openNFTsV4.callStatic.tokenPrice(BigNumber.from(tokenID))).toString()
-    //   );
-    // }
-
-    settingTokenPrice = false;
-    setPriceInput = null;
   };
   /////////////////////////////////////////////////
 </script>
@@ -189,27 +127,17 @@
           </li>
           {#if $nft.price || $nft.price === "0"}
             <li>
-              <!-- <div class="flex"><span class="label">Nft Price</span></div>
+              <div class="flex"><span class="label">Nft Price</span></div>
               <div class="flex">
-                <span class="link overflow-ellipsis" title={$nft.price} target="_blank">
-                  {#if setPriceInput}
-                    <input type="text" bind:value={setPriceInput} disabled={settingTokenPrice} id="set-price-nft" />
-                    <span on:click={setNewTokenPrice} class="btn btn-small btn-outline" title="Confirm token price">
-                      {#if settingTokenPrice}
-                        <i class="fas fa-spinner fa-left c-green refresh" /> wait
-                      {:else}
-                        <i class="fa fa-check" /> Confirm
-                      {/if}
-                    </span>
-                  {:else}
-                    {$nft.price || "0"} Eth
-                    <span on:click={displayPriceInput} class="btn btn-small btn-outline" title="Set price"
-                      ><i class="fa fa-usd" /> Set price</span
-                    >
-                  {/if}
-                </span>
-              </div> -->
-              <NftSetPrice {chainId} {address} {tokenID} nftPrice={$nft.price} />
+                <a
+                  class="link overflow-ellipsis"
+                  href={getEthersConverterLink(chainId, $nft.price)}
+                  title={$nft.price}
+                  target="_blank"
+                >
+                  {$nft.price || "0"} Eth
+                </a>
+              </div>
             </li>
           {/if}
           <!-- {#if $nft.royalties}
@@ -242,10 +170,16 @@
               ><i class="fa fa-code" /><span>Get shortcode</span></a
             >
           {/if}
-          {#if ($nft.price === "0.0" || !$nft.price) && $nft.owner === account}
-            <a href="#transfert-nft-{tokenID}" class="btn btn-small btn-outline" title="Make a gift"
-              ><i class="fa fa-gift" /> Transfer</a
+          {#if $nft.owner === account}
+            <a
+              href="#transfert-nft-{tokenID}"
+              class="btn btn-small btn-outline {$nft.price && $nft.price !== '0.0' ? 'kre-disabled' : ''}"
+              title="Make a gift"
+              aria-disabled={$nft.price && $nft.price !== "0.0"}><i class="fa fa-gift" /> Transfer</a
             >
+          {/if}
+          {#if $nft.owner === account}
+            <NftSell {chainId} {address} {tokenID} nftPrice={$nft.price} />
           {/if}
           {#if $nft.owner !== account}
             <NftBuy {chainId} {address} {tokenID} nftPrice={$nft?.price} />
@@ -307,11 +241,6 @@
   <NftTransfer {chainId} {address} {tokenID} />
 </div>
 
-<!-- Modal buy nft -->
-<!-- <div id="buy-nft-{tokenID}" class="modal-window">
-  <NftBuy {chainId} {address} {tokenID} nftPrice={$nft?.price} />
-</div> -->
-
 <!-- Modal burn nft -->
 <div id="burn-nft-{tokenID}" class="modal-window">
   <NftBurn {chainId} {address} {tokenID} />
@@ -336,5 +265,11 @@
   .btn-burn:hover {
     color: white;
     background: red;
+  }
+
+  a.kre-disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+    pointer-events: none;
   }
 </style>
